requirements.txt
# ── Core ────────────────────────────
requests>=2.31.0
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# ── GitHub ──────────────────────────
PyGithub>=2.2.0

# ── Jira ────────────────────────────
jira>=3.8.0

# ── LLM / Cohere ────────────────────
cohere>=5.0.0

# ── Email ───────────────────────────
sendgrid>=6.11.0

# ── PlantUML parsing ───────────────
# (parseur custom inclus dans le projet)

# ── Templating ──────────────────────
jinja2>=3.1.3

# ── Async ───────────────────────────
asyncio-pool>=0.6.0

# ── CLI ─────────────────────────────
click>=8.1.7
rich>=13.7.0

# ── Webhook Server ──────────────────
fastapi>=0.110.0
uvicorn>=0.27.0

# ── Tests ───────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
respx>=0.20.0
